requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# LLM & ML (OpenAI)
openai==1.12.0
# Optionnel pour fallback local
# torch==2.1.0
# transformers==4.35.0
# peft==0.7.0
# accelerate==0.25.0
# bitsandbytes==0.41.3
# sentencepiece==0.1.99

# Database
supabase==2.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1

# Storage
boto3==1.29.7
minio==7.2.0

# Messaging
kafka-python==2.0.2
aiokafka==0.10.0

# Cache
redis==5.0.1
hiredis==2.2.3

# Data processing
pandas==2.1.3
pyarrow==14.0.1
numpy==1.26.2
scipy==1.11.4

# Validation & Testing
scikit-learn==1.3.2
pytest==7.4.3
pytest-asyncio==0.21.1
# httpx==0.25.2  <-- Laissé libre pour éviter les conflits (géré par openai/supabase)

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
aiofiles==23.2.1
tenacity==8.2.3
cryptography==41.0.7
rich==13.7.0  # Pour la CLI interactive
click==8.1.7  # Pour les arguments CLI

# Monitoring
prometheus-client==0.19.0
structlog==23.2.0
